+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
-- Upgrade and freeze pyright version
 
+## [0.6.5] - 2022-04-18
+- Upgrade and freeze pyright version
 - Rename `compare_version` to `get_max_version` for readability
+- Add user metadata recipe
 
 ## [0.6.4] - 2022-04-11
 - bug fix in `default_create_and_send_custom_email` for emailverification recipe where we were not sending the email if env var was not set.